+++ Yut/Yut/Features/ARGamePlay/ARCoordinator.swift
@@ -1,6 +1,7 @@
 import ARKit
 import RealityKit
 import Combine
+import MultipeerConnectivity
 
 /// ARView의 이벤트를 처리하고 SwiftUI 상태와 연결해주는 총괄 Coordinator
 class ARCoordinator: NSObject, ARSessionDelegate {
@@ -78,7 +79,55 @@ class ARCoordinator: NSObject, ARSessionDelegate {
     }
     
     // MARK: - Game Flow Control
-
+    
+    // '새 게임 준비' 액션을 처리하는 함수
+    func setupNewGame() {
+        Task {
+            // Main 스레드
+            // GameManager 설정, 게임 상태 변경
+            
+            await MainActor.run {
+                // PlayerModel 로드 (PeerID 임시값)
+                let player1 = PlayerModel(name: "노랑", sequence: 1, peerID: MCPeerID(displayName: "Player1"))
+                let player2 = PlayerModel(name: "초록", sequence: 2, peerID: MCPeerID(displayName: "Player2"))
+                
+                guard let arState = self.arState else { return }
+                
+                // GameManager에 실제 플레이어 정보로 새 게임을 설정
+                arState.gameManager.startGame(with: [player1, player2])
+                
+                // PieceManager 윷판 앵커를 알 수 있도록 연결
+                self.pieceManager.boardAnchor = self.boardManager.yutBoardAnchor
+                
+                // 준비 끝, 상태 전환
+                arState.gamePhase = .readyToThrow
+            }
+        }
+    }
+    
+    // 새 말 놓을 때
+    func showDestinationsForNewPiece() {
+        guard let arState = self.arState else { return }
+        let gameManager = arState.gameManager
+        
+        // 시작점에 있는 말과, 윷 결과 게임매니저로부터 가져오기
+        guard let newPiece = gameManager.currentPlayer.pieces.first(where: { $0.position == "_6_6" }),
+              let yutResult = gameManager.yutResult
+        else { return }
+        
+        let destinations = gameManager.routeOptions(for: newPiece, yutResult: yutResult, currentRouteIndex: newPiece.routeIndex)
+        
+        // 목적지 타일 하이라이트
+        if !destinations.isEmpty {
+            let destinationNames = destinations.map { $0.destinationID }
+            self.pieceManager.highlightTiles(named: destinationNames)
+            
+            arState.selectedPiece = newPiece
+            arState.availableDestinations = destinationNames
+            arState.gamePhase = .selectingDestination
+        }
+    }
+    
     // 윷 결과 업데이트 후 -> 움직일 말 선택
     func yutThrowCompleted(with result: YutResult) {
