+++ Yut/Yut/Features/ARGamePlay/Manager /PieceManager.swift
@@ -68,7 +68,7 @@ final class PieceManager {
     
     // MARK: - Highlighting
     
-    // ✨ 1. 기존 함수를 '타일' 하이라이트 전용으로 이름을 명확하게 변경
+    // 하이라이트할 Tile Entity 반환
     func highlightTiles(named tileNames: [String]) {
         guard let boardEntity = boardAnchor?.children.first else {
             print("❌ 윷판 엔티티 없음")
@@ -83,19 +83,18 @@ final class PieceManager {
         }
     }
     
-    // ✨ 2. 제안하신 '말'들을 하이라이트하는 새로운 함수
+    // 하이라이트할 말 Entity 반환
     func highlightMovablePieces(_ pieces: [PieceModel]) {
         clearAllHighlights() // 이전에 있던 모든 하이라이트 제거
         
         for piece in pieces {
-            // PieceModel에서 AR Entity를 직접 가져옵니다.
             guard let pieceEntity = piece.entity as? ModelEntity else { continue }
             applyHighlight(to: pieceEntity)
         }
     }
     
     
-    // ✨ 3. 하이라이트를 적용하는 로직을 별도 함수로 분리 (코드 중복 제거)
+    // 인자로 받은 엔티티의 하이라이트 적용
     private func applyHighlight(to entity: ModelEntity) {
         guard let model = entity.model else { return }
         
@@ -113,37 +112,29 @@ final class PieceManager {
         entity.model = newModel
     }
     
-    // ✨ 4. 모든 하이라이트를 원래대로 되돌리는 함수
+    // 모든 하이라이트를 원래대로 되돌리는 함수
     func clearAllHighlights() {
-        // 딕셔너리에 저장된 모든 엔티티를 순회하며 원래 머티리얼로 복원
         for (entity, originalMaterial) in originalMaterials {
             entity.model?.materials = [originalMaterial]
         }
         originalMaterials.removeAll()
     }
     
-    // ✨ 판 밖에 있던 말을 처음으로 AR 씬에 추가하는 함수
-        func placePieceOnBoard(piece: PieceModel, on tileName: String) {
-            guard let destinationTile = boardAnchor?.findEntity(named: tileName) else {
-                print("❌ 목적지 \(tileName) 타일을 찾을 수 없습니다.")
-                return
-            }
-            
-            
-        
-            
-            
-            
-            
-            let pieceEntity = piece.entity // PlayerModel.load()가 이미 로드해 둔 엔티티
-            
-            pieceEntity.generateCollisionShapes(recursive: true)
-            pieceEntity.scale = [0.3, 8.0, 0.3]
-            pieceEntity.position = [0, 0.2, 0]
-            
-            destinationTile.addChild(pieceEntity)
-
-            
-            print("✅ \(piece.entity.name)을 \(tileName)에 처음으로 배치했습니다.")
+    // 판 밖에 있던 말을 처음으로 AR 씬에 추가하는 함수
+    func placePieceOnBoard(piece: PieceModel, on tileName: String) {
+        guard let destinationTile = boardAnchor?.findEntity(named: tileName) else {
+            print("❌ 목적지 \(tileName) 타일을 찾을 수 없습니다.")
+            return
         }
+        
+        let pieceEntity = piece.entity // PlayerModel.load()가 이미 로드해 둔 엔티티
+        
+        pieceEntity.generateCollisionShapes(recursive: true)
+        pieceEntity.scale = [0.3, 8.0, 0.3]
+        pieceEntity.position = [0, 0.2, 0]
+        
+        destinationTile.addChild(pieceEntity)
+        
+        print("✅ \(piece.entity.name)을 \(tileName)에 처음으로 배치했습니다.")
+    }
 }
